server/requirements.txt
fastapi
uvicorn[standard]
langchain-community
langchain-core
langchain-chroma
langchain-huggingface
chromadb
sentence-transformers
pypdf
python-dotenv
python-multipart
aiofiles
tiktoken
httpx
redis
structlog
prometheus-client
